precheck require optimizations & fixes
