+++ precheck/lib/precheck/rules/all.rb
@@ -1,0 +1,1 @@
+Dir[File.dirname(__FILE__) + '/*.rb'].each { |file| require_relative file }
