+++ precheck/lib/precheck/rules/unreachable_urls_rule.rb
@@ -1,5 +1,7 @@
-require 'precheck/rule'
 require 'addressable'
+require 'faraday_middleware'
+
+require_relative '../rule'
 
 module Precheck
   class UnreachableURLRule < URLRule
